Warn when an HTTP request fails
This would have saved me some time in figuring out that I needed to install
glib-networking (because it's indicated in the reason_phrase that it's required,
but that wasn't being displayed anywhere).  It would also be useful to e.g. a
user who isn't seeing their stocks because their DNS is broken.